+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [UNRELEASED] [paperFromTest]
+## [UNRELEASED] [requirements]
 ### Fixed
 
 - set asynchrony all create components. Then draw only one component per time. Then get data from rest throw web worker.
@@ -29,6 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - new plugin, whom create papers (reports) from template and give link on download and preview
 
 - add set token in single
+
+- check permission for dashboard, main dash page, dash setting
 
 ## [1.44.26] [newSetting]
 ### Adde